chore: prepares to move `semantic-never` to `never` directory
