+++ readme.md
@@ -154,7 +154,7 @@ Looking for something that we don't ship yet? [Raise an issue](https://github.co
 - [`Kind`](https://github.com/ahrjarrett/any-ts/blob/main/src/kind/kind.ts), a simple but powerful [HKT](https://en.wikipedia.org/wiki/Kind_(type_theory)) encoding that supports partial application & re-binding
 - [`integer`](https://github.com/ahrjarrett/any-ts/blob/main/src/number/integer.ts)
 - [`mut`](https://github.com/ahrjarrett/any-ts/blob/main/src/mutable/mutable.ts)
-- [`never`](https://github.com/ahrjarrett/any-ts/blob/main/src/semantic-never/semantic-never.ts), a.k.a. semantic never
+- [`never`](https://github.com/ahrjarrett/any-ts/blob/main/src/never/never.ts), a.k.a. "semantic never"
 - [`nonempty`](https://github.com/ahrjarrett/any-ts/blob/main/src/empty.ts)
 - [`number`](https://github.com/ahrjarrett/any-ts/blob/main/src/number/number.ts)
 - [`bigint`](https://github.com/ahrjarrett/any-ts/blob/main/src/number/bigint.ts)
