+++ src/number/natural.ts
@@ -3,7 +3,7 @@ export { nat }
 import { boolean } from "../boolean/exports"
 import { int } from "./integer"
 import { isPositive } from "./shared"
-import type { never } from "../semantic-never/exports"
+import type { never } from "../never/exports"
 
 namespace nat { export const never: never.not_meant_for_use = void 0 as never }
