+++ src/paths/paths.ts
@@ -4,7 +4,7 @@ export {
 
 import type { any } from "../any/exports"
 import { assert, expect } from "../test/exports"
-import type { never } from "../semantic-never/exports"
+import type { never } from "../never/exports"
 
 type isOptional<key extends keyof type, type>
   = {} extends { [ix in key]: type[ix] } ? true : false
