+++ src/show/associative.ts
@@ -13,7 +13,7 @@ import type { any } from "../any/exports"
 import { assert, describe, expect } from "../test/exports"
 import type { enforce } from "../err/enforce"
 import type { TypeError } from "../err/exports"
-import type { never } from "../semantic-never/exports"
+import type { never } from "../never/exports"
 import type { Universal } from "../universal/exports"
 import type { assoc as _ } from "../associative/exports"
