+++ src/symbol/uri.ts
@@ -1,6 +1,6 @@
-import type { never } from "../semantic-never/exports"
+import type { never } from "../never/exports"
 
 export namespace URI {
-  export const IllegalState: unique symbol = Symbol.for("any-ts/semantic-never::IllegalState")
+  export const IllegalState: unique symbol = Symbol.for("any-ts/never::IllegalState")
   export type IllegalState = typeof never.$$.IllegalState
 }
